+++ Integrations/PhishLabsIOC_EIR/PhishLabsIOC_EIR_test.py
@@ -1,335 +1,324 @@
-import pytest
-
-'''Globals'''
-INDICATORS_EC = [
-    (
-        'url-phishlabs',
-        {
-            "url": "https://google.com/",
-            "malicious": "false",
-            "maliciousDomain": "false"
-        },
-        {
-            'URL': "https://google.com/",
-            'Malicious': "false",
-            'MaliciousDomain': "false"
-        }
-    ),
-    (
-        'attach-phishlabs',
-        {
-            "fileName": "EECO_RFQ__453100Q.pdf",
-            "mimeType": "application/pdf",
-            "md5": "6680e7e593c8286ac51e332d8f72aeec",
-            "sha256": "6b4eeb63b26d3415a529fd95e41e83af19642f2c967b3089a0de8da55c79dd47",
-            "malicious": "false"
-        },
-        {
-            'fileName': "EECO_RFQ__453100Q.pdf",
-            'MimeType': "application/pdf",
-            'MD5': "6680e7e593c8286ac51e332d8f72aeec",
-            'SHA256': "6b4eeb63b26d3415a529fd95e41e83af19642f2c967b3089a0de8da55c79dd47",
-            'Malicious': "false"
-        }
-    ),
-    (
-        'email-ec',
-        {
-            "caseType": "Link",
-            "classification": "No Threat Detected",
-            "subClassification": "No Threat Detected",
-            "severity": "null",
-            "emailReportedBy": "Michael Mammele <support@threatx.com>",
-            "submissionMethod": "Attachment",
-            "sender": "LinkedIn Sales Navigator  <support@threatx.com>",
-            "emailBody": "Test body",
-            "attachments": [],
-            "furtherReviewReason": "null",
-            "offlineUponReview": "false",
-        },
-        {
-            'To': "Michael Mammele <support@threatx.com>",
-            'From': "LinkedIn Sales Navigator  <support@threatx.com>",
-            'Body/HTML': "Test body",
-        }
-    )
-]
-
-INDICATORS_DBOT_EC = [
-    (
-        'url-ec',
-        {
-            "url": "https://google.com/",
-            "malicious": "false",
-            "maliciousDomain": "false"
-        },
-        (
-            {
-                'Indicator': "https://google.com/",
-                'Type': 'URL',
-                'Vendor': "PhishLabs IOC - EIR",
-                'Score': 1
-            },
-            {
-                'Data': "https://google.com/",
-                'Malicious': {
-                    'Vendor': "PhishLabs IOC - EIR",
-                    'Description': "false"
-                }
-            }
-        )
-    ),
-    (
-        'file-ec',
-        {
-            "fileName": "EECO_RFQ__453100Q.pdf",
-            "mimeType": "application/pdf",
-            "md5": "6680e7e593c8286ac51e332d8f72aeec",
-            "sha256": "6b4eeb63b26d3415a529fd95e41e83af19642f2c967b3089a0de8da55c79dd47",
-            "malicious": "false"
-        },
-        (
-            {
-                'Indicator': "EECO_RFQ__453100Q.pdf",
-                'Type': 'File',
-                'Vendor': "PhishLabs IOC - EIR",
-                'Score': 1
-            },
-            {
-                'Name': "EECO_RFQ__453100Q.pdf",
-                'SHA256': "6b4eeb63b26d3415a529fd95e41e83af19642f2c967b3089a0de8da55c79dd47",
-                'MD5': "6680e7e593c8286ac51e332d8f72aeec",
-                'Malicious': {
-                    'Vendor': "PhishLabs IOC - EIR",
-                    'Description': "false"
-                }
-            }
-        )
-    )
-]
+from CommonServerPython import *
 
 
-INDICATORS_TO_LIST_EC = [
-    (
-        'url-ec',
-        [
+def test_create_indicator_no_date():
+    from PhishLabsIOC_EIR import create_indicator_content
+
+    files_json = """
             {
-                "url": "https://google.com/u/gI5Qk",
-                "malicious": 'false',
-                "maliciousDomain": 'false'
-            },
-            {
-                "url": "https://google.com/api/track/v2/5",
-                "malicious": 'false',
-                "maliciousDomain": 'false'
-            },
-            {
-                "url": "https://google.com/",
-                "malicious": 'false',
-                "maliciousDomain": 'false'
-            }
-        ],
-        (
-            [
-                {
-                    'Data': "https://google.com/u/gI5Qk",
-                    'Malicious': {
-                        'Vendor': "PhishLabs IOC - EIR",
-                        'Description': "false"
+                "attributes": [
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "xyz",
+                        "name": "md5",
+                        "value": "c8092abd8d581750c0530fa1fc8d8318"
+                    },
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "abc",
+                        "name": "filetype",
+                        "value": "application/zip"
+                    },
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "qwe",
+                        "name": "name",
+                        "value": "Baycc.zip"
                     }
+                ],
+                "createdAt": "2019-05-14T13:03:45Z",
+                "falsePositive": false,
+                "id": "def",
+                "type": "Attachment",
+                "value": "c8092abd8d581750c0530fa1fc8d8318"
+            } """
+
+    result = {
+        'ID': 'def',
+        'Indicator': 'c8092abd8d581750c0530fa1fc8d8318',
+        'Type': 'Attachment',
+        'CreatedAt': '2019-05-14T13:03:45Z',
+        'UpdatedAt': '',
+        'FalsePositive': False,
+    }
+
+    indicator = json.loads(files_json)
+
+    actual = create_indicator_content(indicator)
+
+    assert actual == result
+
+
+def test_create_indicator_with_none_date():
+    from PhishLabsIOC_EIR import create_indicator_content
+
+    files_json = """
+            {
+                "attributes": [
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "xyz",
+                        "name": "md5",
+                        "value": "c8092abd8d581750c0530fa1fc8d8318"
+                    },
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "abc",
+                        "name": "filetype",
+                        "value": "application/zip"
+                    },
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "qwe",
+                        "name": "name",
+                        "value": "Baycc.zip"
+                    }
+                ],
+                "createdAt": "2019-05-14T13:03:45Z",
+                "updatedAt": "0001-01-01T00:00:00Z",
+                "falsePositive": false,
+                "id": "def",
+                "type": "Attachment",
+                "value": "c8092abd8d581750c0530fa1fc8d8318"
+            } """
+
+    result = {
+        'ID': 'def',
+        'Indicator': 'c8092abd8d581750c0530fa1fc8d8318',
+        'Type': 'Attachment',
+        'CreatedAt': '2019-05-14T13:03:45Z',
+        'UpdatedAt': '',
+        'FalsePositive': False,
+    }
+
+    indicator = json.loads(files_json)
+
+    actual = create_indicator_content(indicator)
+
+    assert actual == result
+
+
+def test_create_indicator_with_date():
+    from PhishLabsIOC_EIR import create_indicator_content
+
+    files_json = """
+            {
+                "attributes": [
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "xyz",
+                        "name": "md5",
+                        "value": "c8092abd8d581750c0530fa1fc8d8318"
+                    },
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "abc",
+                        "name": "filetype",
+                        "value": "application/zip"
+                    },
+                    {
+                        "createdAt": "2019-05-14T13:03:45Z",
+                        "id": "qwe",
+                        "name": "name",
+                        "value": "Baycc.zip"
+                    }
+                ],
+                "createdAt": "2019-05-14T13:03:45Z",
+                "updatedAt": "2019-05-14T13:03:45Z",
+                "falsePositive": false,
+                "id": "def",
+                "type": "Attachment",
+                "value": "c8092abd8d581750c0530fa1fc8d8318"
+            } """
+
+    result = {
+        'ID': 'def',
+        'Indicator': 'c8092abd8d581750c0530fa1fc8d8318',
+        'Type': 'Attachment',
+        'CreatedAt': '2019-05-14T13:03:45Z',
+        'UpdatedAt': '2019-05-14T13:03:45Z',
+        'FalsePositive': False,
+    }
+
+    indicator = json.loads(files_json)
+
+    actual = create_indicator_content(indicator)
+
+    assert actual == result
+
+
+def test_populate_context_files():
+    from PhishLabsIOC_EIR import populate_context, get_file_properties, create_phishlabs_object
+    files_json = """
+        {
+            "attributes": [
+                {
+                    "createdAt": "2019-05-14T13:03:45Z",
+                    "id": "xyz",
+                    "name": "md5",
+                    "value": "c8092abd8d581750c0530fa1fc8d8318"
                 },
                 {
-                    'Data': "https://google.com/api/track/v2/5",
-                    'Malicious': {
-                        'Vendor': "PhishLabs IOC - EIR",
-                        'Description': "false"
-                    }
+                    "createdAt": "2019-05-14T13:03:45Z",
+                    "id": "abc",
+                    "name": "filetype",
+                    "value": "application/zip"
                 },
                 {
-                    'Data': "https://google.com/",
-                    'Malicious': {
-                        'Vendor': "PhishLabs IOC - EIR",
-                        'Description': "false"
-                    }
+                    "createdAt": "2019-05-14T13:03:45Z",
+                    "id": "qwe",
+                    "name": "name",
+                    "value": "Baycc.zip"
                 }
             ],
-            [
-                {
-                    'Indicator': "https://google.com/u/gI5Qk",
-                    'Type': 'URL',
-                    'Vendor': "PhishLabs IOC - EIR",
-                    'Score': 1
-                },
-                {
-                    'Indicator': "https://google.com/api/track/v2/5",
-                    'Type': 'URL',
-                    'Vendor': "PhishLabs IOC - EIR",
-                    'Score': 1
-                },
-                {
-                    'Indicator': "https://google.com/",
-                    'Type': 'URL',
-                    'Vendor': "PhishLabs IOC - EIR",
-                    'Score': 1
-                }
-            ]
-        )
-    )
-]
+            "createdAt": "2019-05-14T13:03:45Z",
+            "falsePositive": false,
+            "id": "def",
+            "type": "Attachment",
+            "updatedAt": "0001-01-01T00:00:00Z",
+            "value": "c8092abd8d581750c0530fa1fc8d8318"
+        } """
+    file = json.loads(files_json)
+    file_md5, file_name, file_type = get_file_properties(file)
 
+    phishlabs_entry = create_phishlabs_object(file)
 
-RAW_RESPONSE_TO_CONTEXT = [
-    [
-        {
-            "id": "INC0660360",
-            "service": "EIR",
-            "title": "Your operating system has been hacked by cybercriminals. Change the authorization method.",
-            "description": "",
-            "status": "Closed",
-            "details": {
-                "caseType": "Response",
-                "classification": "Malicious",
-                "subClassification": "Response - 419 Scam",
-                "severity": "Low",
-                "emailReportedBy": "PhishLabs Phishing Team <not@domain.com>",
-                "submissionMethod": "Attachment",
-                "sender": "<not@domain.com>",
-                "emailBody": "Test",
-                "urls": [
-                    {
-                        "url": "https://google.com/i5/resp",
-                        "malicious": "false",
-                        "maliciousDomain": "false"
-                    }
-                ],
-                "attachments": [
-                    {
-                        "fileName": "EECO_RFQ__453100Q.pdf",
-                        "mimeType": "application/pdf",
-                        "md5": "6680e7e593c8286ac51e332d8f72aeec",
-                        "sha256": "6b4eeb63b26d3415a529fd95e41e83af19642f2c967b3089a0de8da55c79dd47",
-                        "malicious": "false"
-                    }
-                ],
-                "furtherReviewReason": "null",
-                "offlineUponReview": "false"
-            },
-            "created": "2019-11-01T20:55:33Z",
-            "modified": "2019-11-01T21:39:57Z",
-            "closed": "2019-11-01T21:39:57Z",
-            "duration": 2665
-        }
-    ],
-    (
-        [
+    phishlabs_entry['Name'] = file_name
+    phishlabs_entry['Type'] = file_type
+    phishlabs_entry['MD5'] = file_md5
+
+    phishlabs_result = [{
+        'ID': 'def',
+        'CreatedAt': '2019-05-14T13:03:45Z',
+        'Name': 'Baycc.zip',
+        'Type': 'application/zip',
+        'MD5': 'c8092abd8d581750c0530fa1fc8d8318',
+        'Attribute': [
             {
-                'CaseType': "Response",
-                'Classification': "Malicious",
-                'SubClassification': "Response - 419 Scam",
-                'Severity': "Low",
-                'EmailReportedBy': "PhishLabs Phishing Team <not@domain.com>",
-                'SubmissionMethod': "Attachment",
-                'FurtherReviewReason': "null",
-                'ID': "INC0660360",
-                'Title': "Your operating system has been hacked by cybercriminals. Change the authorization method.",
-                'Description': "",
-                'Status': "Closed",
-                'Created': "2019-11-01T20:55:33Z",
-                'Modified': "2019-11-01T21:39:57Z",
-                'Closed': "2019-11-01T21:39:57Z",
-                'Duration': 2665,
-                'Email': {
-                    'EmailBody': "Test",
-                    'Sender': "<not@domain.com>",
-                    'URL': [
-                        {
-                            'URL': "https://google.com/i5/resp",
-                            'Malicious': "false",
-                            'MaliciousDomain': "false"
-                        }
-                    ],
-                    'Attachment': [
-                        {
-                            'fileName': "EECO_RFQ__453100Q.pdf",
-                            'MimeType': "application/pdf",
-                            'MD5': "6680e7e593c8286ac51e332d8f72aeec",
-                            'SHA256': "6b4eeb63b26d3415a529fd95e41e83af19642f2c967b3089a0de8da55c79dd47",
-                            'Malicious': "false"
-                        }
-                    ]
-                }
-            }
-        ],
-        [
-            {
-                'To': "PhishLabs Phishing Team <not@domain.com>",
-                'From': "<not@domain.com>",
-                'Body/HTML': "Test",
-            }
-        ],
-        [
-            {
-                'Name': "EECO_RFQ__453100Q.pdf",
-                'SHA256': "6b4eeb63b26d3415a529fd95e41e83af19642f2c967b3089a0de8da55c79dd47",
-                'MD5': "6680e7e593c8286ac51e332d8f72aeec",
-                'Malicious': {
-                    'Vendor': "PhishLabs IOC - EIR",
-                    'Description': "false"
-                }
-            }
-        ],
-        [
-            {
-                'Data': "https://google.com/i5/resp",
-                'Malicious': {
-                    'Vendor': "PhishLabs IOC - EIR",
-                    'Description': "false"
-                }
-            }
-        ],
-        [
-            {
-                'Indicator': "EECO_RFQ__453100Q.pdf",
-                'Type': 'File',
-                'Vendor': "PhishLabs IOC - EIR",
-                'Score': 1
+                'CreatedAt': '2019-05-14T13:03:45Z',
+                'Type': None,
+                'Name': 'md5',
+                'Value': 'c8092abd8d581750c0530fa1fc8d8318'
             },
             {
-                'Indicator': "https://google.com/i5/resp",
-                'Type': "URL",
-                'Vendor': "PhishLabs IOC - EIR",
-                'Score': 1
+                'CreatedAt': '2019-05-14T13:03:45Z',
+                'Type': None,
+                'Name': 'filetype',
+                'Value': 'application/zip'
+            },
+            {
+                'CreatedAt': '2019-05-14T13:03:45Z',
+                'Type': None,
+                'Name': 'name',
+                'Value': 'Baycc.zip'
             }
         ]
-    )
-]
+    }]
 
-'''Function tests'''
+    global_entry = {
+        'Name': file_name,
+        'Type': file_type,
+        'MD5': file_md5
+    }
+
+    global_result = [{
+        'Name': 'Baycc.zip',
+        'Type': 'application/zip',
+        'MD5': 'c8092abd8d581750c0530fa1fc8d8318'
+    }]
+
+    context = populate_context([], [], [(global_entry, phishlabs_entry)], [])
+
+    assert len(context.keys()) == 2
+    assert context[outputPaths['file']] == global_result
+    assert context['PhishLabs.File(val.ID && val.ID === obj.ID)'] == phishlabs_result
 
 
-class TestHelperFunctions:
-    @pytest.mark.parametrize(argnames='type_ec, test_inputs, test_outputs', argvalues=INDICATORS_EC)
-    def test_indicator_ec(self, type_ec, test_inputs, test_outputs):
-        from PhishLabsIOC_EIR import indicator_ec
-        result = indicator_ec(indicator=test_inputs,
-                              type_ec=type_ec)
-        assert result == test_outputs
+def test_populate_context_emails():
+    from PhishLabsIOC_EIR import populate_context, get_email_properties, create_phishlabs_object
+    emails_json = """
+        {
+           "attributes":[
+              {
+                 "createdAt":"2019-05-13T16:54:18Z",
+                 "id":"abc",
+                 "name":"email-body",
+                 "value":"-----Original Message-----From: A Sent: Monday, May 13, 2019 12:22 PMTo:"
+              },
+              {
+                 "createdAt":"2019-05-13T16:54:18Z",
+                 "id":"def",
+                 "name":"from",
+                 "value":"someuser@contoso.com"
+              },
+              {
+                 "createdAt":"2019-05-13T16:54:18Z",
+                 "id":"cf3182ca-92ec-43b6-8aaa-429802a99fe5",
+                 "name":"to",
+                 "value":"example@gmail.com"
+              }
+           ],
+           "createdAt":"2019-05-13T16:54:18Z",
+           "falsePositive":false,
+           "id":"ghi",
+           "type":"E-mail",
+           "updatedAt":"0001-01-01T00:00:00Z",
+           "value":"FW: Task"
+        } """
+    email = json.loads(emails_json)
+    email_body, email_to, email_from = get_email_properties(email)
 
-    @pytest.mark.parametrize(argnames='type_ec, test_inputs, test_outputs', argvalues=INDICATORS_DBOT_EC)
-    def test_indicator_dbot_ec(self, type_ec, test_inputs, test_outputs):
-        from PhishLabsIOC_EIR import indicator_dbot_ec
-        result = indicator_dbot_ec(indicator=test_inputs,
-                                   type_ec=type_ec)
-        assert result == test_outputs
+    phishlabs_entry = create_phishlabs_object(email)
 
-    @pytest.mark.parametrize(argnames='type_ec, test_inputs, test_outputs', argvalues=INDICATORS_TO_LIST_EC)
-    def test_indicators_to_list_ec(self, type_ec, test_inputs, test_outputs):
-        from PhishLabsIOC_EIR import indicators_to_list_ec
-        result = indicators_to_list_ec(indicators=test_inputs,
-                                       type_ec=type_ec)
-        assert result == test_outputs
+    phishlabs_entry['To'] = email_to,
+    phishlabs_entry['From'] = email_from,
+    phishlabs_entry['Body'] = email_body
+    phishlabs_entry['Subject'] = email.get('value')
 
-    def test_raw_response_to_context(self):
-        from PhishLabsIOC_EIR import raw_response_to_context
-        result = raw_response_to_context(incidents=RAW_RESPONSE_TO_CONTEXT[0])
-        assert result == RAW_RESPONSE_TO_CONTEXT[1]
+    phishlabs_result = [{
+        'ID': 'ghi',
+        'CreatedAt': '2019-05-13T16:54:18Z',
+        'To': ('example@gmail.com',),
+        'From': ('someuser@contoso.com',),
+        'Body': '-----Original Message-----From: A Sent: Monday, May 13, 2019 12:22 PMTo:',
+        'Subject': 'FW: Task',
+        'Attribute':
+            [{
+                'CreatedAt': '2019-05-13T16:54:18Z',
+                'Type': None,
+                'Name': 'email-body',
+                'Value': '-----Original Message-----From: A Sent: Monday, May 13, 2019 12:22 PMTo:'
+            },
+            {
+                'CreatedAt': '2019-05-13T16:54:18Z',
+                'Type': None,
+                'Name': 'from',
+                'Value': 'someuser@contoso.com'
+            },
+            {
+                'CreatedAt': '2019-05-13T16:54:18Z',
+                'Type': None,
+                'Name': 'to',
+                'Value': 'example@gmail.com'
+            }]
+    }]
+
+    global_entry = {
+        'To': email_to,
+        'From': email_from,
+        'Body': email_body,
+        'Subject': email.get('value')
+    }
+
+    global_result = [{
+        'To': 'example@gmail.com',
+        'From': 'someuser@contoso.com',
+        'Body': '-----Original Message-----From: A Sent: Monday, May 13, 2019 12:22 PMTo:',
+        'Subject': 'FW: Task'
+    }]
+
+    context = populate_context([], [], [], [], [(global_entry, phishlabs_entry)])
+
+    assert len(context.keys()) == 2
+    assert context['Email'] == global_result
+    assert context['PhishLabs.Email(val.ID && val.ID === obj.ID)'] == phishlabs_result